+++ ArgumentParser.Tests/Arguments.cs
@@ -1,0 +1,9 @@
+﻿using ArgumentParser.Attributes;
+
+namespace ArgumentParser.Tests
+{
+    [ArgumentClass]
+    public class Arguments
+    {        
+    }
+}
