+++ ArgumentParser.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.30717.126
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ArgumentParser", "ArgumentParser\ArgumentParser.csproj", "{0E31E1CC-599F-455B-9204-B53C0F0F7140}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ArgumentParser.Tests", "ArgumentParser.Tests\ArgumentParser.Tests.csproj", "{D6629B30-47FE-43FD-BFBA-CAF544F1789B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{0E31E1CC-599F-455B-9204-B53C0F0F7140}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0E31E1CC-599F-455B-9204-B53C0F0F7140}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0E31E1CC-599F-455B-9204-B53C0F0F7140}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D6629B30-47FE-43FD-BFBA-CAF544F1789B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D6629B30-47FE-43FD-BFBA-CAF544F1789B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D6629B30-47FE-43FD-BFBA-CAF544F1789B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D6629B30-47FE-43FD-BFBA-CAF544F1789B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
